Organized folder structure in utils folder
