+++ modules/auth/components/Avatar.tsx
@@ -1,6 +1,6 @@
 import { NextPage } from "next";
 import React, { useState, useEffect } from "react";
-import { supabase } from "../../../utils/supabaseClient";
+import { supabase } from "../../../utils/db/supabaseClient";
 import { Avatar } from "../interfaces/Avatar";
 import Image from "next/image";
