+++ pages/_app.tsx
@@ -1,5 +1,6 @@
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
+import { useEffect } from 'react'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return <Component {...pageProps} />
