+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from "next";
 import { useState, useEffect } from "react";
-import { supabase } from "../utils/supabaseClient";
+import { supabase } from "../utils/db/supabaseClient";
 import HomePage from "../components/Home";
 import Account from "../modules/auth/components/Account";
 import { Session } from "@supabase/supabase-js";
